+++ src/main/java/subway/ui/line/FindSingleLineController.java
@@ -1,4 +1,4 @@
-package subway.ui;
+package subway.ui.line;
 
 import java.sql.SQLException;
 import org.springframework.http.ResponseEntity;
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import subway.application.line.FindSingleLineUseCase;
+import subway.application.line.usecase.FindSingleLineUseCase;
 import subway.ui.dto.response.LineResponse;
 
 @RestController
