+++ src/main/java/subway/ui/station/AttachFrontStationController.java
@@ -1,4 +1,4 @@
-package subway.ui;
+package subway.ui.station;
 
 import java.sql.SQLException;
 import org.springframework.http.HttpStatus;
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import subway.application.station.AttachFrontStationUseCase;
+import subway.application.station.usecase.AttachFrontStationUseCase;
 import subway.ui.dto.request.AttachStationRequest;
 
 @RestController
